only 1st/3rd classification tasks

+++ datasets/charades_ego_plus_charades.py
@@ -23,7 +23,7 @@ class CharadesMeta(Charades):
                      'firsttime_neg': 0,
                      'n': 0,
                      'n_ego': 0,
-                     'class_target': torch.zeros(157),
+                     'class_target': torch.zeros(157).int(),
                      })
         if self.split == 'val_video':
             return ims, target, meta
